Move TermsTrie and SuffixTrie into their own modules
